refactor: remove barrel file from shared
