+++ src/app/features/pokemon/application/services/get-pokemon-list/get-pokemon-list.ts
@@ -5,7 +5,7 @@ import { type ServiceCommand } from '@/core/domain/command/service-command'
 import { error, success } from '@/core/domain/either/either'
 import type { PokemonListItem } from '@/app/features/pokemon/domain/models'
 import type { Paginated, PaginationParams } from '@/core/domain/types/paginated'
-import type { PaginatedDTO } from '@/app/shared/types'
+import type { PaginatedDTO } from '@/app/shared/types/paginated-dto'
 import {
   pokemonListMapper,
   type PokemonListItemDTO
