+++ src/app/features/pokemon/presentation/components/PokemonNotFound.tsx
@@ -1,4 +1,4 @@
-import { useTranslation } from '@/app/shared/hooks'
+import { useTranslation } from '@/app/shared/hooks/useTranslation'
 
 export const PokemonNotFound = () => {
   const { t } = useTranslation('common')
