+++ src/app/features/pokemon/presentation/components/SkeletonPokeCard.tsx
@@ -1,4 +1,4 @@
-import { Skeleton } from '@/app/shared/components/ui'
+import { Skeleton } from '@/app/shared/components/ui/skeleton'
 
 export const SkeletonPokeCard = () => (
   <div className='bg-blue-light w-full rounded-2xl px-4 pt-4 pb-2 shadow-[1px_1px_2px_0_rgba(0,0,0,0.1),_-1px_-1px_2px_0_rgba(0,0,0,0.05)] md:max-w-80'>
