+++ src/app/features/pokemon/presentation/screens/PokemonSearchScreen.tsx
@@ -1,7 +1,7 @@
 import { Link, useParams } from '@tanstack/react-router'
 import { PokeCard } from '../components'
-import { ArrowIcon } from '@/app/shared/components/icons'
-import { Button } from '@/app/shared/components/ui'
+import { ArrowIcon } from '@/app/shared/components/icons/arrow'
+import { Button } from '@/app/shared/components/ui/button'
 import { useTranslation } from 'react-i18next'
 
 const PokemonSearchScreen = () => {
