--- src/app/shared/components/index.ts
@@ -1,2 +1,0 @@
-export { LanguageSwitcher } from './LanguageSwitcher'
-export { ResponsiveDialog } from './ResponsiveDialog'
