--- src/app/shared/components/ui/index.ts
@@ -1,9 +1,0 @@
-export { Button } from './button'
-export { Toaster } from './sonner'
-export { Skeleton } from './skeleton'
-export { Badge } from './badge'
-export {
-  Collapsible,
-  CollapsibleContent,
-  CollapsibleTrigger
-} from './collapsible'
