--- src/app/shared/hooks/index.ts
@@ -1,2 +1,0 @@
-export * from './useTranslation'
-export * from './useMidiaQuery'
