--- src/app/shared/types/index.ts
@@ -1,1 +1,0 @@
-export * from './paginated-dto'
